+++ constants/icons.ts
@@ -11,6 +11,7 @@ import CheckIcon from "@/assets/icons/check.svg";
 import ChevronLeftIcon from "@/assets/icons/chevron-left.svg";
 import ChevronRightIcon from "@/assets/icons/chevron-right.svg";
 import CommentIcon from "@/assets/icons/comment.svg";
+import EmailIcon from "@/assets/icons/email.svg";
 import EyeOffIcon from "@/assets/icons/eye-off.svg";
 import EyeIcon from "@/assets/icons/eye.svg";
 import FaceDefaultIcon from "@/assets/icons/face_default.svg";
@@ -31,6 +32,8 @@ import HomeFilledIcon from "@/assets/icons/home_fill.svg";
 import KakaoIcon from "@/assets/icons/kakao.svg";
 import KebabMenuIcon from "@/assets/icons/kebab.svg";
 import MeatballIcon from "@/assets/icons/meatball.svg";
+import OAuthApple from "@/assets/icons/oauth_apple.svg";
+import OAuthGoogle from "@/assets/icons/oauth_google.svg";
 import PlusIcon from "@/assets/icons/plus.svg";
 import ProfileIcon from "@/assets/icons/profile.svg";
 import ProfileFilledIcon from "@/assets/icons/profile_fill.svg";
@@ -39,6 +42,7 @@ import SettingIcon from "@/assets/icons/setting.svg";
 import ThreadIcon from "@/assets/icons/thread.svg";
 import TrashCanIcon from "@/assets/icons/trash_can.svg";
 import XIcon from "@/assets/icons/x.svg";
+import SignInKoKKoK from "@assets/icons/signin_kokkok.svg";
 
 export default {
   AppleIcon,
@@ -54,6 +58,7 @@ export default {
   ChevronRightIcon,
   CommentIcon,
   CheckIcon,
+  EmailIcon,
   EyeIcon,
   EyeOffIcon,
   FaceDefaultIcon,
@@ -74,11 +79,14 @@ export default {
   KakaoIcon,
   KebabMenuIcon,
   MeatballIcon,
+  OAuthApple,
+  OAuthGoogle,
   PlusIcon,
   ProfileIcon,
   ProfileFilledIcon,
   SearchIcon,
   SettingIcon,
+  SignInKoKKoK,
   TrashCanIcon,
   ThreadIcon,
   XIcon,
